refactor: ReservationUser 삭제, Reservation 수정에 따른 연관관계 수정 #3
